Improve VPM-API
Use a simple classic Sinatra app
